+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/request/ExaminationRequestDTO.java
@@ -1,0 +1,26 @@
+package com.ivanovp.medical_record.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ExaminationRequestDTO {
+
+    @NotNull(message = "Patient ID is required")
+    private Long patientId;
+
+    private Long diagnosisId;
+
+    private String treatment;
+
+    @NotNull(message = "Price is required")
+    @Positive(message = "Price must be positive")
+    private BigDecimal price;
+}